requirements.txt
python-dotenv 
pandas 
numpy
pymongo 
certifi
pymongo[srv]==4.11
scikit-learn
# -e . add this at the last for for packaging okay , it refer to the setup.py file